Fix: import should not break on Blender 4.3 and below (#19)
This PR addresses compatibility issues with using the script on versions of Blender older than 4.4. The main adaptation introduced in this patch is a guard against the Blender API version code for access to the `ActionSlots` API, which was initially introduced in Blender 4.4. On versions older than 4.4, `action.fcurves` is accessed directly, as that is the lowest level of granularity possible with those APIs.

Reviewed-on: https://codeberg.org/leftmostcat/nu-blender/pulls/19
Reviewed-by: leftmostcat <leftmostcat@noreply.codeberg.org>
Co-authored-by: Isabelle Santin <isabelle@chosenfewsoftware.com>
Co-committed-by: Isabelle Santin <isabelle@chosenfewsoftware.com>